% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorsets.R
\name{colorsets}
\alias{colorsets}
\title{colorsets}
\usage{
colorsets(col = "total")
}
\arguments{
\item{col}{= color sub-theme in my theme that users want to look their display and hex codes .
The defaults shows the total colors in my theme.}
}
\description{
Function shows display and hex codes of my color theme. My theme contains 7 sub-theme including
formal, main, cool, hot, mixed and grey.
}
\examples{
colorsets()
colorsets("formal")
}
