% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{conf}
\alias{conf}
\title{Conf}
\usage{
conf(x, se, alpha = 0.05, df, two = TRUE)
}
\arguments{
\item{x}{= value(mean)}

\item{se}{=  standard error}

\item{alpha}{= confidence level}

\item{df}{= degree of freedom}

\item{two}{= TRUE if it is a two-tailed test and FALSE it is a one-tailed test}
}
\description{
calculate confidence interval
}
\examples{
conf(20, 1.5, alpha = 0.025, df = 500, two = FALSE)
}
