% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstat.R
\name{fstat}
\alias{fstat}
\title{fstat}
\usage{
fstat(sser, sseu, j, n, k)
}
\arguments{
\item{sser}{SSE of restricted value}

\item{sseu}{SSE of unrestricted value}

\item{j}{hypothesis test}

\item{n}{total observation}

\item{k}{number of coefficent}
}
\description{
fstat
}
\examples{
fvalue(sser = 200,sseu = 150,j = 2,n = 320,k =5)
}
