% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots}
\alias{plots}
\title{plots}
\usage{
plots(date, serie, dffit)
}
\arguments{
\item{date}{serie}

\item{serie}{observed data}

\item{dffit}{datafrome containing the fitted valued from different models}
}
\description{
This function allows user to plot the fitted value or the forecast from diffrent model compared to the observed data.
The graphs apply ggplot2, so the user can add other option which applied to ggplot to the result from this function
such as ggtitle.
}
\examples{
Using economics dataset from ggplot
#prepare data
pop = ts(economics$pop, start = c(1967,7),freq = 12)
date = seq(1967.50,2015.25,length=length(pop))
dffit = data.frame(
fitm1  = tslm(pop~trend)$fitted,
fitm2  = exp(tslm(log(pop) ~ trend)$fitted))

#apply function
re = plots(date = date,serie = pop, dffit = dffit)
re +  ggtitle("Time series Trend graph using differnet Models")
}
