% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_parn.R
\name{scale_color_parn}
\alias{scale_color_parn}
\title{scale_color_parn}
\usage{
scale_color_parn(palette = "formal", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{my color sub-theme in my theme (use colorsets function to explore more details)}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not
(i.e. if the variable is not discrete, we should apply color gradient scale so discrete = FALSE)}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{col}{color sub-theme in my theme that users want to look their display and hex codes .
The defaults shows the total colors in my theme.}
}
\value{

}
\description{
Color theme function is used with graph created by ggplots. This function is apply for ggplot2 with color scale function.
(ggplot with fill function, use another one called scale_fill_parn). My color theme has contains 7 sub-theme including formal, main, cool, hot, mixed and grey.
}
\examples{
data(iris)
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
geom_point(size = 4) +
scale_color_parn(palette = "formal")

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
geom_point(size = 4, alpha = .6) +
scale_color_parn(discrete = FALSE, palette = "mixed")
}
