% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_parn.R
\name{scale_fill_parn}
\alias{scale_fill_parn}
\title{scale_fill_parn}
\usage{
scale_fill_parn(palette = "formal", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{my color sub-theme in my theme (use colorsets function to explore more details)}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not
(i.e. if the variable is not discrete, we should apply color gradient scale so discrete = FALSE)}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{col}{color sub-theme in my theme that users want to look their display and hex codes .
The defaults shows the total colors in my theme.}
}
\value{

}
\description{
Color theme function is used with graph created by ggplots. This function is apply for ggplot2 with color scale function.
(ggplot with color function, use another one called scale_color_parn). My color theme has contains 7 sub-theme including formal, main, cool, hot, mixed and grey.
}
\examples{
data(iris)

ggplot(mpg, aes(y = class)) + geom_bar(aes(fill = drv), position = position_stack(reverse = TRUE), alpha = 0.8) + scale_fill_parn("hot")
ggplot(mtcars, aes(factor(cyl), fill = factor(vs))) + geom_bar() + scale_fill_parn("formal")
}
