% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigt.R
\name{sigt}
\alias{sigt}
\title{sigt}
\usage{
sigt(x, a, se, df, alpha = 0.05, two = TRUE)
}
\arguments{
\item{x}{observed value}

\item{a}{hyptosis value}

\item{se}{standard error}

\item{df}{degree of freedom}

\item{alpha}{confidence level}

\item{two}{= TRUE if it is a two-tailed test and FALSE it is a one-tailed test}
}
\description{
calculate t statistics given variable and show criticals t-statistic.
}
\examples{
sigt(100,99,1.5,600)


}
